split functions to multiple go files
