+++ api/testing/testing.go
@@ -1,0 +1,7 @@
+package testing
+
+import "lmm/api/db"
+
+func init() {
+	db.SetDefaultDatabaseName("lmm_test")
+}
